+++ neo/game/gamesys/Event.h
@@ -104,7 +104,7 @@ public:
 								~idEvent();
 
 	static idEvent				*Alloc( const idEventDef *evdef, int numargs, va_list args );
-	static void					CopyArgs( const idEventDef *evdef, int numargs, va_list args, int data[ D_EVENT_MAXARGS ]  );
+	static void					CopyArgs( const idEventDef *evdef, int numargs, va_list args, intptr_t data[ D_EVENT_MAXARGS ]  );
 
 	void						Free( void );
 	void						Schedule( idClass *object, const idTypeInfo *cls, int time );
